pyproject.toml
[project]
name = "google-tunix"
version = "0.1.6"
authors = [
  { name = "Tunix Developers", email = "tunix-dev@google.com" },
]
description = "A lightweight JAX-native LLM post-training framework."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Apache-2.0" }
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "datasets",
  "flax>=0.11.1",
  "fsspec", # notebook dependency
  "google-metrax>=0.2.3",
  "grain",
  "huggingface_hub",
  "jaxtyping",
  "jinja2", # Huggingface chat template
  "kagglehub",
  "numba",
  "omegaconf", # CLI config
  "pylatexenc", # Eval result parsing
  "python-dotenv", # Huggingface API key
  "qwix",
  "sentencepiece",
  "sympy", # Eval result parsing
  "tensorflow_datasets",
  "tqdm",
  "transformers",
  "hf_transfer", # Huggingface caching in CI
  "joblib>=1.5.2",
  "draccus>=0.11.5",
  "jax[tpu]>=0.8.0",
  "tpu-info>=0.7.1",
  "kaggle>=1.8.2",
]

[project.optional-dependencies]
docs = [
    "sphinx>=8.2.3",
    "sphinx-book-theme>=1.1.4",  # Older versions fail to pin pydata-sphinx-theme
    "sphinx-autodoc-typehints",
    "ipython>=8.8.0",  # 8.7.0 has ipython3 lexer error
    "myst-nb>=1.3.0",
    "matplotlib>=3.10.0",
    "sphinx-gallery>=0.19.0",
    "sphinx-collections>=0.0.1",
    "sphinx_contributors",
]
prod = [
  "jax[tpu]>=0.6.0,!=0.7.2", # Jax 0.7.2 has performance regression on OSS
]
dev = [
  # Manully install vLLM & tpu-inferece, which depends on jax[tpu]==0.7.2
]

[project.urls]
Source = "https://github.com/google/tunix"

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tunix*", "research*"]

[tool.pyink]
line-length = 80
unstable = true
target-version = []
pyink-indentation = 2
pyink-use-majority-quotes = true
pyink-annotation-pragmas = [
  "noqa",
  "pylint:",
  "type: ignore",
  "pytype:",
  "mypy:",
  "pyright:",
  "pyre-",
  "@title",
  "@param",
  "@markdown",
  "@tidy",
  "@test",
]
pyink-ipynb-indentation = 2
pyink-ipynb-unicode-escape = true

[tool.isort]
profile = "google"       # or "black", etc.
known_third_party = ["tunix"]
